Use rust brotli implementation
brotli contains some large data tables,
and servo already uses the rust brotli crate,
so removing the vendored brotli allows us to reduce
the binary size of servo.

This requires us to create a new crate for woff2, since it
depends on brotli decompressor functions and we need to
specify this dependency, otherwise there will be linker errors.
