+++ Cargo.toml
@@ -8,6 +8,7 @@ license = "BSD-3-Clause"
 edition = "2015"
 
 [dependencies]
+fontsan-woff2 = { version = "0.1", path = "src/deps/woff2"}
 libc = "0.2"
 libz-rs-sys = {  version = "0.5.1", features = ["export-symbols"], optional = true }
 libz-sys = { version = "1.1", default-features = false, features = ["libc" ], optional = true}
