+++ src/lib.rs
@@ -3,6 +3,7 @@
 // Use of this source code is governed by a BSD-style license
 // that can be found in the LICENSE file.
 
+extern crate fontsan_woff2;
 extern crate libc;
 #[cfg(feature = "libz-rs")]
 extern crate libz_rs_sys;
